added extra two section

+++ src/Components/Carts/Cart.jsx
@@ -1,17 +1,48 @@
 /* eslint-disable react/prop-types */
 
+import Swal from "sweetalert2";
+
 
 const Cart = ({phone}) => {
 
     const {_id,name,price,image} = phone
 
     const handleDeleteCartItem = () => {
+
+      Swal.fire({
+        title: 'Are you sure?',
+        text: "You won't be able to revert this!",
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#3085d6',
+        cancelButtonColor: '#d33',
+        confirmButtonText: 'Yes, delete it!'
+      }).then((result) => {
+        if (result.isConfirmed) {
+          fetch(`http://localhost:5000/phones/mycart/delete/${_id}`, {
+            method: "DELETE",
+          })
+          .then(response=> response.json())
+          .then(data=>{
+           
+            if (data.deletedCount>0) {
+              Swal.fire(
+                'Deleted!',
+                'Your file has been deleted.',
+                'success'
+              )
+            }
+          })
+          .then(()=>{
+            location.reload();
+          })
+          
+        }
+      })
+
+
    
-        fetch(`http://localhost:5000/phones/mycart/delete/${_id}`, {
-          method: "DELETE",
-        })
-        .then(response=> response.json())
-        .then(data=>console.log(data))
+
